add an extra load for non-new but unloaded
Made an improvement to the "deferred" / "load_only" set of strategy options
where if a certain object is loaded from two different logical paths within
one query, attributes that have been configured by at least one of the
options to be populated will be populated in all cases, even if other load
paths for that same object did not set this option. previously, it was
based on randomness as to which "path" addressed the object first.

Fixes: #8166
Change-Id: I923a1484721d3a04d490ef882bc9fa609c9cd077