test transfer of default, insert_default
right now "default" goes to Column.default unconditionally
if insert_default is not present, including if dataclasses
are in use where the field effectively now does two things.
This generally works out because Python side default
can be assigned to the object or picked up by Core in any case.
However, we might want to look into later on migrating this
to have the fields act more separately.  I think it's
"OK" for now, will try to doc that this might change.

Change-Id: I30f2085ec79a6464da4d5c578500848c70d55ec2