Add alembic extra to setup.cfg
This merge request adds a new `extras_require` to setup.cfg to allow users to install Alembic to be installed with SQLAlchemy via extras requires.

While alembic currently requires SQLAlchemy to function, it does not allow the user to include the database driver or mypy at the same time. By providing this option, a user will be able to install SQLAlchemy, a database driver, Mypy, and Alembic via a single command for easy inclusion into projects.

Thank you for your consideration of this change.